Functionally works, need to fix all tests
