+++ tests/test_client.py
@@ -1,6 +1,6 @@
 import pytest
 
-from aiospamc.client import Client
+from aiospamc.client import Client2
 from aiospamc.responses import (
     CantCreateException,
     ConfigException,
